reddit_stats/requirements.txt
praw
asyncpg
discord.py
